Implemented Collection Methods for Discovery service
